Read.me
This application is a Weather API. It works by using HTML,CSS, and JavaScript. It starts by taking your GEOLocation onload(). If it is blocked it will send
an alert and say to either allow location or simply use the search bar to see weather in a city specified by user. The goal was to make a modern and simplistic
design that is pleasing to the eyes.